[Bronze II] Title: Hashing, Time: 128 ms, Memory: 14196 KB -BaekjoonHub

+++ 백준/Bronze/15829. Hashing/Hashing.java
@@ -6,10 +6,10 @@ class Main {
     public static void main(String[] args) throws IOException{
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
-        int N = Integer.parseInt(br.readLine());
+        long N = Integer.parseInt(br.readLine());
         String str = br.readLine();
 
-        int total = 0;
+        long total = 0;
 
         for (int i = 0; i < str.length(); i++) {
             total += Math.pow(31, i) * (str.charAt(i) - 'a' + 1);
